Update AIShipCaptain.java
Instead of immediately retreating, when a dangerous amount of incoming missiles is detected, the targetted AI-stack will first consider whether the missiles will be able to hit it when it moves backwards. If not, it will do so and only retreat if the missiles hitting the stack are unavoidable.

When missiles would run out before reaching the stack, it will not retreat at all.