Update RacesIntelligenceUI.java
Techs which you could steal but are obsolete to your empire are now displayed in grey on the intelligence-tab in the races-screen.

When you select your own empire on the intelligence-tab in the races-screen the techs you are lacking now also are color-coded in grey, yellow or orange depending on whether they are obsolete, stealable or not stealable.