New option for the Smart spending tools: Refit and balance
New option for the Smart spending tools to maintain a balance between industry and ecology to maximise efficiency, this by matching the number of factories to the number of workers. When a robotic upgrade is available, the growth/factory balance will continue until the colony is fully populated, at which point the factories will be upgraded.