Colony Ships in wars
Reverted a previous change that prevented the AI from building colony-ships when it had fewer military than their enemies.

Rationale: While those colony ships seem like a waste of resources, the wars are not so much about direct fights anyways. What is important is being able to reach the enemies' colonies. So a small but constant flow of colonizers can be necessary to get access deeper into the enemies' territory.