Autoplay Fix and attack-target-scoring-adjustment
Fixed an issue that caused the AI not to attack with player-designed ships when handing control over to AI in an ongoing playthrough.

The scoring for what colonies to attack no longer takes the value-ratio of population to missile-bases into account. This change means you can no longer simply build a missile-base to force the AI to ignore the closest colonies.