Speed-requirement for missiles in ship-designs reduced
When considering whether to use missiles for their ship-designs the AI now has a lower speed-requirement for the missiles based on the enemie's designs speeds.
The distance-advantage of repulsors is split above the two turns the missile will travel instead being considered twice.
It is also taken into consideration that the missile only needs to get into 0.7 range instead of having to catch up to the target's center.