Fixed several issues that lead to Sakkra- and Bulrathi-AI not Transport-rushing other empires in the early-game when they should
Population currently flying around in transports is now considered when evaluating the aggressiveness of Bulrathi- or Sakkra-like empires.

Invasions are now considered for all enemy-empires and not only the ones that we are already at war with.

For the calculation of an empires' power-level the population that is currently flying around in transports is now taken into consideration too.