More colony-ships for everyone
The AI will no longer be as conservative about building colony-ships once they have the techs needed for going to war.
This was meant for them to not overbuild early on but they limited themselves too much later on so someone who would just stay out of conflicts while having a lo of colony-ships could more easily take over everything for themselves.