AI smarter about Blackhole-generator and Stasis-field-usage
When a defending fleet has black-hole-generators or stasis-field-generators they will now no longer try to concede the first-hit to cloaked attackers. That is because black-hole-generator success-chance isn't reduced by cloaking-device and stasis-field has 100% chance to succeed.

AI-ships equipped with stasis-field-generator will now select targets as if they could kill them in order to make it more likely to freeze the most dangerous stack.